introduce base class TestClientApi
