+++ utils/docker_fixtures/_test_clients/_core.py
@@ -1,15 +1,26 @@
 import os
-from pathlib import Path
 import shutil
 import subprocess
+import time
+import urllib.parse
+from http import HTTPStatus
+from pathlib import Path
 from typing import TextIO
 
 from docker.models.images import Image
+from docker.models.containers import Container
 import pytest
+import requests
+from _pytest.outcomes import Failed
 
 from utils._logger import logger
-
 from utils.docker_fixtures._core import get_docker_client
+
+
+def _fail(message: str):
+    """Used to mak a test as failed"""
+    logger.error(message)
+    raise Failed(message, pytrace=False) from None
 
 
 class TestClientFactory:
@@ -129,3 +140,60 @@ class TestClientFactory:
         log_path.parent.mkdir(parents=True, exist_ok=True)
 
         return log_path.open("w+", encoding="utf-8")
+
+
+class TestClientApi:
+    """Base Test Client class for parametric and integration_framework scenario"""
+
+    container: Container
+    timeout: int
+    _session: requests.Session
+    _base_url: str
+
+    def __init__(self, url: str, timeout: int, container: Container):
+        self._base_url = url
+        self._session = requests.Session()
+        self.container = container
+        self.timeout = timeout
+
+        # wait for server to start
+        self._wait(timeout)
+
+    def container_restart(self):
+        self.container.restart()
+        self._wait(self.timeout)
+
+    def _wait(self, timeout: float):
+        delay = 0.01
+        for _ in range(int(timeout / delay)):
+            try:
+                if self._is_alive():
+                    break
+            except Exception:
+                if self.container.status != "running":
+                    self._print_logs()
+                    message = f"Container {self.container.name} status is {self.container.status}. Please check logs."
+                    _fail(message)
+            time.sleep(delay)
+        else:
+            self._print_logs()
+            message = f"Timeout of {timeout} seconds exceeded waiting for HTTP server to start. Please check logs."
+            _fail(message)
+
+    def _url(self, path: str) -> str:
+        return urllib.parse.urljoin(self._base_url, path)
+
+    def _is_alive(self) -> bool:
+        self.container.reload()
+        return (
+            self.container.status == "running"
+            and self._session.get(self._url("/non-existent-endpoint-to-ping-until-the-server-starts")).status_code
+            == HTTPStatus.NOT_FOUND
+        )
+
+    def _print_logs(self):
+        try:
+            logs = self.container.logs().decode("utf-8")
+            logger.debug(f"Logs from container {self.container.name}:\n\n{logs}")
+        except Exception:
+            logger.error(f"Failed to get logs from container {self.container.name}")
