+++ utils/docker_fixtures/_test_clients/_test_client_framework_integrations.py
@@ -1,19 +1,13 @@
 from collections.abc import Generator
 import contextlib
-from http import HTTPStatus
-import time
-import urllib.parse
 
-from _pytest.outcomes import Failed
-from docker.models.containers import Container
 import pytest
 import requests
 
-from utils._logger import logger
 
 from utils.docker_fixtures._core import docker_run, get_host_port
 from utils.docker_fixtures._test_agent import TestAgentAPI
-from ._core import TestClientFactory
+from ._core import TestClientFactory, TestClientApi
 
 
 class FrameworkTestClientFactory(TestClientFactory):
@@ -89,54 +83,10 @@ class FrameworkTestClientFactory(TestClientFactory):
         )
 
 
-def _fail(message: str):
-    """Used to mak a test as failed"""
-    logger.error(message)
-    raise Failed(message, pytrace=False) from None
-
-
-class FrameworkTestClientApi:
+class FrameworkTestClientApi(TestClientApi):
     """API to interact with the tracer+framework server running in a docker container for
     INTEGRATIONS_FRAMEWORK scenarios.
     """
-
-    def __init__(self, url: str, timeout: int, container: Container):
-        self._base_url = url
-        self._session = requests.Session()
-        self.container = container
-        self.timeout = timeout
-
-        # wait for server to start
-        self._wait(timeout)
-
-    def container_restart(self):
-        self.container.restart()
-        self._wait(self.timeout)
-
-    def _wait(self, timeout: float):
-        delay = 0.01
-        for _ in range(int(timeout / delay)):
-            try:
-                if self._is_alive():
-                    break
-            except Exception:
-                if self.container.status != "running":
-                    self._print_logs()
-                    message = f"Container {self.container.name} status is {self.container.status}. Please check logs."
-                    _fail(message)
-            time.sleep(delay)
-        else:
-            self._print_logs()
-            message = f"Timeout of {timeout} seconds exceeded waiting for HTTP server to start. Please check logs."
-            _fail(message)
-
-    def _is_alive(self) -> bool:
-        self.container.reload()
-        return (
-            self.container.status == "running"
-            and self._session.get(self._url("/non-existent-endpoint-to-ping-until-the-server-starts")).status_code
-            == HTTPStatus.NOT_FOUND
-        )
 
     def request(
         self, method: str, url: str, body: dict | None = None, *, raise_for_status: bool = True
@@ -145,13 +95,3 @@ class FrameworkTestClientApi:
         if raise_for_status:
             resp.raise_for_status()
         return resp
-
-    def _print_logs(self):
-        try:
-            logs = self.container.logs().decode("utf-8")
-            logger.debug(f"Logs from container {self.container.name}:\n\n{logs}")
-        except Exception:
-            logger.error(f"Failed to get logs from container {self.container.name}")
-
-    def _url(self, path: str) -> str:
-        return urllib.parse.urljoin(self._base_url, path)
